Fix bug that waterways were still removed
The waterways were still removed since `NetworkCleaner` removes all
nodes that are not in the biggest cluster of the network. Since the
roads and waterways are two separate clusters only the roads were kept.
This is fixed by adding a `num` parameter to the `run()` function of
`NetworkCleaner` that sets the number of biggest clusters that should be
kept and setting it to two.